Summary: Merge upstream code

Testing: CICD

Reviewers: kuaiwei, yuleil

Issue: https://github.com/dragonwell-project/dragonwell21/issues/1

+++ src/hotspot/os/linux/cgroupV1Subsystem_linux.cpp
@@ -320,7 +320,7 @@ jlong CgroupV1MemoryController::cache_usage_in_bytes() {
     return OSCONTAINER_ERROR;
   }
   log_trace(os, container)("Cache usage is: " JULONG_FORMAT, cache);
-  return cache;
+  return (jlong)cache;
 }
 
 jlong CgroupV1MemoryController::kernel_memory_usage_in_bytes() {
